+++ talaiot/src/main/kotlin/com/cdsap/talaiot/provider/PublishersProvider.kt
@@ -43,7 +43,6 @@ class PublishersProvider(
                     InfluxDbPublisher(
                         this,
                         logger,
-                        SimpleRequest(logger),
                         executor
                     )
                 )
