+++ talaiot/src/main/kotlin/com/cdsap/talaiot/publisher/InfluxDbPublisher.kt
@@ -7,6 +7,7 @@ import com.cdsap.talaiot.request.Request
 import okhttp3.OkHttpClient
 import org.influxdb.InfluxDB
 import org.influxdb.InfluxDBFactory
+import org.influxdb.InfluxDBIOException
 import org.influxdb.dto.BatchPoints
 import org.influxdb.dto.Point
 import java.util.concurrent.Executor
@@ -27,10 +28,6 @@ class InfluxDbPublisher(
      */
     private val logTracker: LogTracker,
     /**
-     * Interface to send the measurements to an external service
-     */
-    private val requestPublisher: Request,
-    /**
      * Executor to schedule a task in Background
      */
     private val executor: Executor
@@ -43,36 +40,42 @@ class InfluxDbPublisher(
 
         if (influxDbPublisherConfiguration.url.isEmpty() ||
             influxDbPublisherConfiguration.dbName.isEmpty() ||
-            influxDbPublisherConfiguration.taskMetricName.isEmpty()
+            influxDbPublisherConfiguration.taskMetricName.isEmpty() ||
+            influxDbPublisherConfiguration.buildMetricName.isEmpty()
         ) {
             println(
-                "InfluxDbPublisher not executed. Configuration requires url, dbName and urlMetrics: \n" +
+                "InfluxDbPublisher not executed. Configuration requires url, dbName, taskMetricName and buildMetricName: \n" +
                         "influxDbPublisher {\n" +
                         "            dbName = \"tracking\"\n" +
                         "            url = \"http://localhost:8086\"\n" +
-                        "            taskMetricName = \"tracking\"\n" +
+                        "            buildMetricName = \"build\"\n" +
+                        "            taskMetricName = \"task\"\n" +
                         "}\n" +
                         "Please update your configuration"
             )
             return
         }
 
-        val _db = createDb()
+        try {
+            val _db = createDb()
 
-        val measurements = createTaskPoints(report)
-        val buildMeasurement = createBuildPoint(report)
+            val measurements = createTaskPoints(report)
+            val buildMeasurement = createBuildPoint(report)
 
-        if (!measurements.isNullOrEmpty()) {
-            val points = BatchPoints.builder()
-                .points(measurements)
-                .point(buildMeasurement)
-                .build()
+            if (!measurements.isNullOrEmpty()) {
+                val points = BatchPoints.builder()
+                    .points(measurements)
+                    .point(buildMeasurement)
+                    .build()
 
-            executor.execute {
-                _db.write(points)
+                executor.execute {
+                    _db.write(points)
+                }
+            } else {
+                logTracker.log("Empty content")
             }
-        } else {
-            logTracker.log("Empty content")
+        } catch (e: InfluxDBIOException){
+            logTracker.log("InfluxDb Error: ${e.message}")
         }
     }
