+++ talaiot/src/test/kotlin/com/cdsap/talaiot/e2e/InfluxDbPublisherNotProperlyDefinedBuildTest.kt
@@ -1,13 +1,12 @@
-package com.cdsap.talaiot.functional
+package com.cdsap.talaiot.e2e
 
 import io.kotlintest.specs.BehaviorSpec
 import org.gradle.testkit.runner.GradleRunner
-import org.gradle.testkit.runner.TaskOutcome
 
-class NoOutputsBuildTest : BehaviorSpec({
+class InfluxDbPublisherNotProperlyDefinedBuildTest : BehaviorSpec({
     given("Build Gradle File") {
         val testProjectDir = TemporaryFolder()
-        `when`("Talaiot is included but no logger mode included") {
+        `when`("Talaiot is included with InfluxDbPublisher but missing the url") {
             testProjectDir.create()
             val buildFile = testProjectDir.newFile("build.gradle")
             buildFile.appendText(
@@ -19,7 +18,10 @@ class NoOutputsBuildTest : BehaviorSpec({
 
                   talaiot{
                     publishers {
-                      outputPublisher {}
+                      influxDbPublisher {
+                           dbName = "tracking"
+                           taskMetricName = "tracking"
+                      }
                   }
                }
             """
@@ -29,11 +31,8 @@ class NoOutputsBuildTest : BehaviorSpec({
                 .withArguments("assemble")
                 .withPluginClasspath()
                 .build()
-            then("no logs are shown in the output") {
-                assert(!result.output.contains("OutputPublisher"))
-                assert(!result.output.contains("¯\\_(ツ)_/¯"))
-                assert(result.task(":assemble")?.outcome == TaskOutcome.SUCCESS)
-
+            then("logs displays the InfluxDbPublisher error") {
+                assert(result.output.contains("InfluxDbPublisher not executed. Configuration requires url, dbName, taskMetricName and buildMetricName"))
             }
             testProjectDir.delete()
         }
