+++ talaiot/src/test/kotlin/com/cdsap/talaiot/e2e/NoOutputBuildTest.kt
@@ -1,13 +1,13 @@
-package com.cdsap.talaiot.functional
+package com.cdsap.talaiot.e2e
 
 import io.kotlintest.specs.BehaviorSpec
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
 
-class PushGatewayPublisherBuildTest : BehaviorSpec({
+class NoOutputsBuildTest : BehaviorSpec({
     given("Build Gradle File") {
         val testProjectDir = TemporaryFolder()
-        `when`("Talaiot is included with PushGatewayPublisher") {
+        `when`("Talaiot is included but no logger mode included") {
             testProjectDir.create()
             val buildFile = testProjectDir.newFile("build.gradle")
             buildFile.appendText(
@@ -18,14 +18,10 @@ class PushGatewayPublisherBuildTest : BehaviorSpec({
                    }
 
                   talaiot{
-                      logger = com.cdsap.talaiot.logger.LogTracker.Mode.INFO
-                      publishers {
-                         pushGatewayPublisher {
-                             url = "http://localhost:9091"
-                             taskJobName = "tracking"
-                    }
-                }
-            }
+                    publishers {
+                      outputPublisher {}
+                  }
+               }
             """
             )
             val result = GradleRunner.create()
@@ -33,10 +29,9 @@ class PushGatewayPublisherBuildTest : BehaviorSpec({
                 .withArguments("assemble")
                 .withPluginClasspath()
                 .build()
-            then("logs are shown in the output and including the pushGateway format") {
-                println(result.output)
-                assert(result.output.contains("PushGatewayPublisher"))
-                assert(result.output.contains(":assemble{state=\"EXECUTED\","))
+            then("no logs are shown in the output") {
+                assert(!result.output.contains("OutputPublisher"))
+                assert(!result.output.contains("¯\\_(ツ)_/¯"))
                 assert(result.task(":assemble")?.outcome == TaskOutcome.SUCCESS)
 
             }
