+++ talaiot/src/test/kotlin/com/cdsap/talaiot/e2e/OutputPublisherBuildTest.kt
@@ -1,12 +1,13 @@
-package com.cdsap.talaiot.functional
+package com.cdsap.talaiot.e2e
 
 import io.kotlintest.specs.BehaviorSpec
 import org.gradle.testkit.runner.GradleRunner
+import org.gradle.testkit.runner.TaskOutcome
 
-class InfluxDbPublisherNotProperlyDefinedBuildTest : BehaviorSpec({
+class OutputPublisherBuildTest : BehaviorSpec({
     given("Build Gradle File") {
         val testProjectDir = TemporaryFolder()
-        `when`("Talaiot is included with InfluxDbPublisher but missing the url") {
+        `when`("Talaiot is included with OutputPublisher") {
             testProjectDir.create()
             val buildFile = testProjectDir.newFile("build.gradle")
             buildFile.appendText(
@@ -17,11 +18,9 @@ class InfluxDbPublisherNotProperlyDefinedBuildTest : BehaviorSpec({
                    }
 
                   talaiot{
-                    publishers {
-                      influxDbPublisher {
-                           dbName = "tracking"
-                           taskMetricName = "tracking"
-                      }
+                    logger = com.cdsap.talaiot.logger.LogTracker.Mode.INFO
+                     publishers {
+                      outputPublisher {}
                   }
                }
             """
@@ -31,8 +30,11 @@ class InfluxDbPublisherNotProperlyDefinedBuildTest : BehaviorSpec({
                 .withArguments("assemble")
                 .withPluginClasspath()
                 .build()
-            then("logs displays the InfluxDbPublisher error") {
-                assert(result.output.contains("InfluxDbPublisher not executed. Configuration requires url, dbName and urlMetrics"))
+            then("logs are shown in the output with the shrugged") {
+                assert(result.output.contains("OutputPublisher"))
+                assert(result.output.contains("¯\\_(ツ)_/¯"))
+                assert(result.task(":assemble")?.outcome == TaskOutcome.SUCCESS)
+
             }
             testProjectDir.delete()
         }
