+++ talaiot/src/test/kotlin/com/cdsap/talaiot/e2e/PushGatewayPublisherBuildTest.kt
@@ -1,15 +1,13 @@
-package com.cdsap.talaiot.functional
+package com.cdsap.talaiot.e2e
 
 import io.kotlintest.specs.BehaviorSpec
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
-import java.io.File
 
-class DependencyGraphPublisherTest : BehaviorSpec({
-
+class PushGatewayPublisherBuildTest : BehaviorSpec({
     given("Build Gradle File") {
         val testProjectDir = TemporaryFolder()
-        `when`("Talaiot is included with TaskDependencyGraph") {
+        `when`("Talaiot is included with PushGatewayPublisher") {
             testProjectDir.create()
             val buildFile = testProjectDir.newFile("build.gradle")
             buildFile.appendText(
@@ -19,26 +17,28 @@ class DependencyGraphPublisherTest : BehaviorSpec({
                       id 'talaiot'
                    }
 
-                  talaiot {
-                    publishers {
-                      taskDependencyGraphPublisher {
-                          html = true
-                          gexf = true
-                      }
+                  talaiot{
+                      logger = com.cdsap.talaiot.logger.LogTracker.Mode.INFO
+                      publishers {
+                         pushGatewayPublisher {
+                             url = "http://localhost:9091"
+                             taskJobName = "tracking"
                     }
-                  }
+                }
+            }
             """
             )
-
             val result = GradleRunner.create()
                 .withProjectDir(testProjectDir.getRoot())
                 .withArguments("assemble")
                 .withPluginClasspath()
                 .build()
-            then("html and gexf files are generated") {
-                assert(File("${testProjectDir.getRoot()}/build/reports/talaiot/taskgraph/htmlTaskDependency.html").exists())
-                assert(File("${testProjectDir.getRoot()}/build/reports/talaiot/taskgraph/gexfTaskDependency.gexf").exists())
+            then("logs are shown in the output and including the pushGateway format") {
+                println(result.output)
+                assert(result.output.contains("PushGatewayPublisher"))
+                assert(result.output.contains(":assemble{state=\"EXECUTED\","))
                 assert(result.task(":assemble")?.outcome == TaskOutcome.SUCCESS)
+
             }
             testProjectDir.delete()
         }
